third_party/blink/web_tests/fast/dom/row-inner-text.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<TABLE ID="tbl" CELLSPACING="1" CELLPADDING="1" BORDER="1">
	<TR id="r">
		<TH>(1,1) 
		</TH>
		<TH>(1,2) 
		</TH>
		<TH>(1,3) 
		</TH>
		<TH>(1,4) 
		</TH>
		<TH>(1,5) 
		</TH>
	</TR>
</TABLE>
<script>
test(() => {
  assert_equals(document.getElementById('r').innerText,
                '(1,1)\t(1,2)\t(1,3)\t(1,4)\t(1,5)');
}, 'TR.innerText');
</script>
